#362 rename: Catogory 관련 엔티티들을 category 패키지로 이동
