+++ src/main/java/com/floney/floney/book/repository/BookLineRepositoryImpl.java
@@ -6,6 +6,7 @@ import com.floney.floney.analyze.dto.request.AnalyzeRequestByBudget;
 import com.floney.floney.analyze.dto.response.AnalyzeResponseByCategory;
 import com.floney.floney.analyze.dto.response.QAnalyzeResponseByCategory;
 import com.floney.floney.book.domain.entity.*;
+import com.floney.floney.book.domain.entity.category.Category;
 import com.floney.floney.book.dto.process.*;
 import com.floney.floney.book.dto.request.AllOutcomesRequest;
 import com.floney.floney.book.util.DateFactory;
@@ -27,7 +28,6 @@ import static com.floney.floney.book.domain.entity.QBook.book;
 import static com.floney.floney.book.domain.entity.QBookLine.bookLine;
 import static com.floney.floney.book.domain.entity.QBookLineCategory.bookLineCategory;
 import static com.floney.floney.book.domain.entity.QBookUser.bookUser;
-import static com.floney.floney.book.domain.entity.QCategory.category;
 import static com.floney.floney.book.dto.constant.AssetType.*;
 import static com.floney.floney.common.constant.Status.ACTIVE;
 import static com.floney.floney.common.constant.Status.INACTIVE;
