+++ src/test/java/com/floney/floney/book/CategoryFixture.java
@@ -1,6 +1,9 @@
 package com.floney.floney.book;
 
 import com.floney.floney.book.domain.entity.*;
+import com.floney.floney.book.domain.entity.category.Category;
+import com.floney.floney.book.domain.entity.category.DefaultCategory;
+import com.floney.floney.book.domain.entity.category.RootCategory;
 import com.floney.floney.book.dto.request.CreateCategoryRequest;
 import com.floney.floney.book.dto.response.CreateCategoryResponse;
 import com.floney.floney.book.domain.entity.category.BookCategory;
@@ -15,43 +18,43 @@ public class CategoryFixture {
 
     public static CreateCategoryRequest createRootRequest() {
         return CreateCategoryRequest.builder()
-                .bookKey(BOOK_KEY)
-                .name(ROOT)
-                .build();
+            .bookKey(BOOK_KEY)
+            .name(ROOT)
+            .build();
     }
 
     public static RootCategory createDefaultRoot(String name) {
         return RootCategory.builder()
-                .name(name)
-                .build();
+            .name(name)
+            .build();
     }
 
     public static DefaultCategory createDefaultChild(Category parent, String name) {
         return DefaultCategory.builder()
-                .parent(parent)
-                .name(name)
-                .build();
+            .parent(parent)
+            .name(name)
+            .build();
     }
 
     public static BookCategory createRootCategory(Book book) {
         return BookCategory.builder()
-                .book(book)
-                .name(ROOT)
-                .status(ACTIVE)
-                .build();
+            .book(book)
+            .name(ROOT)
+            .status(ACTIVE)
+            .build();
     }
 
     public static DefaultCategory incomeBookCategory() {
         return DefaultCategory.builder()
-                .name("수입")
-                .build();
+            .name("수입")
+            .build();
     }
 
 
     public static DefaultCategory outComeBookCategory() {
         return DefaultCategory.builder()
-                .name("지출")
-                .build();
+            .name("지출")
+            .build();
     }
 
 
@@ -71,32 +74,32 @@ public class CategoryFixture {
 
     public static BookCategory createChildCategory(Category parent, Book book) {
         return BookCategory.builder()
-                .book(book)
-                .parent(parent)
-                .name(CHILD)
-                .status(ACTIVE)
-                .build();
+            .book(book)
+            .parent(parent)
+            .name(CHILD)
+            .status(ACTIVE)
+            .build();
     }
 
 
     public static CreateCategoryRequest createBookCategory() {
         return CreateCategoryRequest.builder()
-                .bookKey(BOOK_KEY)
-                .parent(ROOT)
-                .name(CHILD)
-                .build();
+            .bookKey(BOOK_KEY)
+            .parent(ROOT)
+            .name(CHILD)
+            .build();
     }
 
     public static CreateCategoryResponse categoryRootResponse() {
         return CreateCategoryResponse.builder()
-                .name(ROOT)
-                .build();
+            .name(ROOT)
+            .build();
     }
 
     public static CreateCategoryResponse categoryChildResponse() {
         return CreateCategoryResponse.builder()
-                .name(CHILD)
-                .build();
+            .name(CHILD)
+            .build();
     }
 
 }
