+++ src/test/java/com/floney/floney/book/CategoryServiceTest.java
@@ -2,7 +2,7 @@ package com.floney.floney.book;
 
 import com.floney.floney.book.dto.request.CreateCategoryRequest;
 import com.floney.floney.book.domain.entity.Book;
-import com.floney.floney.book.domain.entity.RootCategory;
+import com.floney.floney.book.domain.entity.category.RootCategory;
 import com.floney.floney.book.domain.entity.category.BookCategory;
 import com.floney.floney.book.repository.BookRepository;
 import com.floney.floney.book.repository.category.CategoryRepository;
@@ -46,15 +46,15 @@ public class CategoryServiceTest {
         Book savedBook = BookFixture.createBookWith("1234");
 
         given(bookRepository.findBookByBookKeyAndStatus(BOOK_KEY, Status.ACTIVE))
-                .willReturn(ofNullable(savedBook));
+            .willReturn(ofNullable(savedBook));
 
         BookCategory rootCategory = createRootCategory(savedBook);
 
         given(categoryRepository.save(any(BookCategory.class)))
-                .willReturn(rootCategory);
+            .willReturn(rootCategory);
 
         Assertions.assertThat(categoryService.createUserCategory(request))
-                .isEqualTo(categoryRootResponse());
+            .isEqualTo(categoryRootResponse());
 
     }
 
@@ -65,13 +65,13 @@ public class CategoryServiceTest {
         RootCategory rootCategory = createDefaultRoot("ROOT");
 
         given(categoryRepository.findParentCategory(ROOT))
-                .willReturn(ofNullable(rootCategory));
+            .willReturn(ofNullable(rootCategory));
 
         given(bookRepository.findBookByBookKeyAndStatus(BOOK_KEY, Status.ACTIVE))
-                .willReturn(ofNullable(savedBook));
+            .willReturn(ofNullable(savedBook));
 
         given(categoryRepository.save(any(BookCategory.class)))
-                .willReturn(createChildCategory(rootCategory, savedBook));
+            .willReturn(createChildCategory(rootCategory, savedBook));
 
         CreateCategoryRequest request = createBookCategory();
         Assertions.assertThat(categoryService.createUserCategory(request)).isEqualTo(categoryChildResponse());
@@ -83,24 +83,24 @@ public class CategoryServiceTest {
         RootCategory rootCategory = createDefaultRoot("ROOT");
 
         given(categoryRepository.findParentCategory(ROOT))
-                .willReturn(ofNullable(rootCategory));
+            .willReturn(ofNullable(rootCategory));
 
         given(bookRepository.findBookByBookKeyAndStatus(BOOK_KEY, Status.ACTIVE))
-                .willReturn(Optional.empty());
+            .willReturn(Optional.empty());
 
         CreateCategoryRequest request = createBookCategory();
         Assertions.assertThatThrownBy(() -> categoryService.createUserCategory(request))
-                .isInstanceOf(NotFoundBookException.class);
+            .isInstanceOf(NotFoundBookException.class);
     }
 
     @Test
     @DisplayName("루트 카테고리를 찾을 수 없으면 NotFoundCategoryException을 발생한다")
     void not_found_root() {
         given(categoryRepository.findParentCategory(ROOT))
-                .willReturn(Optional.empty());
+            .willReturn(Optional.empty());
 
         CreateCategoryRequest request = createBookCategory();
         Assertions.assertThatThrownBy(() -> categoryService.createUserCategory(request))
-                .isInstanceOf(NotFoundCategoryException.class);
+            .isInstanceOf(NotFoundCategoryException.class);
     }
 }
